Move invoices and invoices_datatables to invoices folder. Move customers and customers_datatables to customers folder.
